+++ BUGames2014_Squirrel/ActionState.cpp
@@ -120,14 +120,14 @@ void ActionState::DoCollisions() {
 		if ((*iter)->GetState() == PlayerState::Falling) {
 
 			//...check to see if they touch ANY platform block in the level...
-			for (int i = 0; i < theLevel->GetPlatformBlocks()->size(); i++) {
+			for (unsigned int i = 0; i < theLevel->GetPlatformBlocks()->size(); i++) {
 				if (collision->boxCollision(
-					(*iter)->getX(),
-					(*iter)->getY(), 
+					(int)(*iter)->getX(),
+					(int)(*iter)->getY(), 
 					(*iter)->getW(), 
 					(*iter)->getH(),
-					theLevel->GetPlatformBlocks()->at(i)->getX(), 
-					theLevel->GetPlatformBlocks()->at(i)->getY(), 
+					(int)theLevel->GetPlatformBlocks()->at(i)->getX(), 
+					(int)theLevel->GetPlatformBlocks()->at(i)->getY(), 
 					theLevel->GetPlatformBlocks()->at(i)->getW(), 
 					theLevel->GetPlatformBlocks()->at(i)->getH())) {
 
@@ -144,14 +144,14 @@ void ActionState::DoCollisions() {
 			bool noHit = true;
 
 			//...but check them against every platform...
-			for (int i = 0; i < theLevel->GetPlatformBlocks()->size(); i++) {
+			for (unsigned int i = 0; i < theLevel->GetPlatformBlocks()->size(); i++) {
 				if (collision->boxCollision(
-					(*iter)->getX(),
-					(*iter)->getY(), 
+					(int)(*iter)->getX(),
+					(int)(*iter)->getY(), 
 					(*iter)->getW(), 
 					(*iter)->getH(),
-					theLevel->GetPlatformBlocks()->at(i)->getX(), 
-					theLevel->GetPlatformBlocks()->at(i)->getY(), 
+					(int)theLevel->GetPlatformBlocks()->at(i)->getX(), 
+					(int)theLevel->GetPlatformBlocks()->at(i)->getY(), 
 					theLevel->GetPlatformBlocks()->at(i)->getW(), 
 					theLevel->GetPlatformBlocks()->at(i)->getH())) {
 
@@ -193,7 +193,7 @@ void ActionState::DoPlayerAttacks() {
 	std::vector<CharacterBase*>& players = characterManager->GetPlayers();
 
 	//for every player in the players list...
-	for (int i = 0; i < players.size(); i++) {
+	for (unsigned int i = 0; i < players.size(); i++) {
 
 		//...if they are inactive, continue on to the next
 		if (!players.at(i)->GetIsActive()) continue;
@@ -202,19 +202,19 @@ void ActionState::DoPlayerAttacks() {
 		if (input->WasPadButtonPressed(players.at(i)->GetPadID(), SDL_CONTROLLER_BUTTON_B)) {
 			
 			//...iterate over the other players, check for collisions between them
-			for (int j = 0; j < players.size(); j++) {
+			for (unsigned int j = 0; j < players.size(); j++) {
 				
 				//don't check against self
 				if (j == i) continue;
 				if (!players.at(j)->GetIsActive()) continue;
 
 				if (collision->boxCollision(
-					players.at(i)->getX(),
-					players.at(i)->getY(),
+					(int)players.at(i)->getX(),
+					(int)players.at(i)->getY(),
 					players.at(i)->getW(),
 					players.at(i)->getH(),
-					players.at(j)->getX(),
-					players.at(j)->getY(),
+					(int)players.at(j)->getX(),
+					(int)players.at(j)->getY(),
 					players.at(j)->getW(),
 					players.at(j)->getH())) {
